requirements.txt
Django~=2.2.1
djangorestframework~=3.9.3
Pillow~=6.0.0
